game player was moved to single component
